requirement.txt
alembic==1.11.2
aniso8601==8.0.0
annotated-types==0.5.0
anyio==3.7.1
apispec==6.3.0
asgiref==3.7.1
attrs==23.1.0
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cinemagoer==2023.5.1
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cors==1.0.1
customtkinter==4.6.1
darkdetect==0.8.0
distlib==0.3.6
Django==4.0.10
django-jazzmin==2.6.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
et-xmlfile==1.1.0
face-recognition-models==0.3.0
fastapi==0.103.0
filelock==3.12.0
flasgger==0.9.7.1
Flask==2.3.2
flask-babel==4.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.8
flask-smorest==0.42.0
Flask-SQLAlchemy==2.4.3
flask-swagger-ui==4.11.1
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
gTTS==2.3.2
h11==0.14.0
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.0
idna==3.4
IMDbPY==2022.7.9
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mistune==3.0.1
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pipenv==2023.5.19
platformdirs==3.5.1
psycopg2==2.9.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.7.0
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rpds-py==0.9.2
six==1.15.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
tldextract==3.4.4
typing_extensions==4.7.1
tzdata==2023.3
unicorn==2.0.1.post1
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
webargs==8.3.0
websocket-client==1.1.0
Werkzeug==2.3.7
zope.event==5.0
zope.interface==6.0
